Backend Concluded

+++ src/main/java/br/com/wk/tech/blooddonorsapi/dto/BloodTypePotentialDonorsDTO.java
@@ -9,5 +9,5 @@ import lombok.*;
 @Builder
 public class BloodTypePotentialDonorsDTO {
     private String bloodType;
-    private Double numberOfPotentialDonors;
+    private Integer numberOfPotentialDonors;
 }
